refactor(chat): split chat-service into focused modules
- chat-prompt.ts: system prompt template loading + building
- chat-mcp.ts: MCP credential issuance + client connection
- chat-history.ts: message history reconstruction
- chat-service.ts: lean orchestrator importing from modules
- All modules export centralized objects (chatPrompt, chatMcp, chatHistory)
